On branch kitkat-base  Your branch is ahead of 'origin/main' by 1 commit.    (use "git push" to publish your local commits)
 Changes to be committed:
	modified:   cst-kitkat/cts
	modified:   suite/audio_quality/BUILD.txt
	deleted:    suite/audio_quality/client/Android.mk
	modified:   suite/audio_quality/client/src/com/android/cts/audiotest/AudioProtocol.java
	modified:   suite/audio_quality/client/src/com/android/cts/audiotest/CtsAudioClientActivity.java
	modified:   suite/audio_quality/executable/src/main.cpp
	modified:   suite/audio_quality/lib/include/FileUtil.h
	modified:   suite/audio_quality/lib/include/GenericFactory.h
	modified:   suite/audio_quality/lib/include/Log.h
	modified:   suite/audio_quality/lib/include/StringUtil.h
	modified:   suite/audio_quality/lib/include/audio/AudioHardware.h
	modified:   suite/audio_quality/lib/include/audio/AudioProtocol.h
	modified:   suite/audio_quality/lib/include/audio/AudioRemote.h
	modified:   suite/audio_quality/lib/include/audio/Buffer.h
	modified:   suite/audio_quality/lib/include/audio/RemoteAudio.h
	modified:   suite/audio_quality/lib/include/task/ModelBuilder.h
	modified:   suite/audio_quality/lib/include/task/TaskCase.h
	modified:   suite/audio_quality/lib/include/task/TaskProcess.h
	modified:   suite/audio_quality/lib/src/BuiltinProcessing.cpp
	modified:   suite/audio_quality/lib/src/FileUtil.cpp
	modified:   suite/audio_quality/lib/src/GenericFactory.cpp
	modified:   suite/audio_quality/lib/src/Log.cpp
	modified:   suite/audio_quality/lib/src/RWBuffer.h
	modified:   suite/audio_quality/lib/src/SignalProcessingImpl.cpp
	modified:   suite/audio_quality/lib/src/StringUtil.cpp
	modified:   suite/audio_quality/lib/src/audio/AudioHardware.cpp
	modified:   suite/audio_quality/lib/src/audio/AudioProtocol.cpp
	modified:   suite/audio_quality/lib/src/audio/AudioRemote.cpp
	modified:   suite/audio_quality/lib/src/audio/AudioSignalFactory.cpp
	modified:   suite/audio_quality/lib/src/audio/Buffer.cpp
	modified:   suite/audio_quality/lib/src/audio/RemoteAudio.cpp
	modified:   suite/audio_quality/lib/src/task/ModelBuilder.cpp
	modified:   suite/audio_quality/lib/src/task/TaskCase.cpp
	modified:   suite/audio_quality/lib/src/task/TaskGeneric.cpp
	modified:   suite/audio_quality/lib/src/task/TaskInput.cpp
	modified:   suite/audio_quality/lib/src/task/TaskProcess.cpp
	modified:   suite/audio_quality/lib/src/task/TaskSave.cpp
	modified:   suite/audio_quality/lib/src/task/TaskSound.cpp
	modified:   suite/audio_quality/meson.build
	modified:   suite/audio_quality/test/AudioRecordPlayTestCommon.h
	modified:   suite/audio_quality/test/BufferTest.cpp
	modified:   suite/audio_quality/test/ClientInterfaceTest.cpp
	modified:   suite/audio_quality/test/LogTest.cpp
	modified:   suite/audio_quality/test/RemoteAudioFakeTcpTest.cpp
	modified:   suite/audio_quality/test/SignalProcessingInterfaceTest.cpp
	modified:   suite/audio_quality/test/StringUtilTest.cpp
	modified:   suite/audio_quality/test/TaskCaseTest.cpp
	modified:   suite/audio_quality/test/TaskSequentialTest.cpp
	modified:   suite/audio_quality/test_description/conf/check_conf.py
	modified:   suite/audio_quality/test_description/conf/detect_usb_audio.py
	modified:   suite/audio_quality/test_description/processing/calc_delay.py
	modified:   suite/audio_quality/test_description/processing/calc_thd.py
	modified:   suite/audio_quality/test_description/processing/check_spectrum.py
	modified:   suite/audio_quality/test_description/processing/check_spectrum_playback.py
	modified:   suite/audio_quality/test_description/processing/consts.py
	modified:   suite/audio_quality/test_description/processing/example.py
	modified:   suite/audio_quality/test_description/processing/gen_random.py
	modified:   suite/audio_quality/test_description/processing/playback_thd.py
	modified:   suite/audio_quality/test_description/processing/recording_thd.py
	modified:   suite/audio_quality/test_description/processing_main.py
	modified:   tools/tradefed-host/.project
	modified:   tools/tradefed-host/tests/.project

 Changes not staged for commit:
	modified:   cst-kitkat/cts (modified content)

 Untracked files:
	.idea/
	apps/cts-usb-accessory/meson.build
	suite/audio_quality.backup/
	suite/audio_quality/client/.gradle/
	suite/audio_quality/client/README_ANDROID_BUILD.md
	suite/audio_quality/client/build.gradle
	suite/audio_quality/client/build/
	suite/audio_quality/client/gradlew
	suite/audio_quality/client/settings.gradle
	suite/audio_quality/compat/include/tinyalsa/
	suite/audio_quality/meson/
	suite/audio_quality/reports/

+++ suite/audio_quality/test/LogTest.cpp
@@ -48,15 +48,32 @@ TEST_F(LogTest, logTest) {
     int64_t d = 3;
     int64_t e = 4;
     int64_t f = 5;
-    printf("printf %lld %lld %lld %lld %lld %lld\n", a, b, c, d, e, f);
-    LOGD(  "logd   %lld %lld %lld %lld %lld %lld", a, b, c, d, e, f);
-    LOGV(  "logv   %lld %lld %lld %lld %lld %lld", a, b, c, d, e, f);
-    LOGI(  "logi   %lld %lld %lld %lld %lld %lld", a, b, c, d, e, f);
-    LOGW(  "logw   %lld %lld %lld %lld %lld %lld", a, b, c, d, e, f);
-    LOGE(  "loge   %lld %lld %lld %lld %lld %lld", a, b, c, d, e, f);
+    printf("printf %lld %lld %lld %lld %lld %lld\n",
+            static_cast<long long>(a), static_cast<long long>(b),
+            static_cast<long long>(c), static_cast<long long>(d),
+            static_cast<long long>(e), static_cast<long long>(f));
+    LOGD("logd   %lld %lld %lld %lld %lld %lld",
+            static_cast<long long>(a), static_cast<long long>(b),
+            static_cast<long long>(c), static_cast<long long>(d),
+            static_cast<long long>(e), static_cast<long long>(f));
+    LOGV("logv   %lld %lld %lld %lld %lld %lld",
+            static_cast<long long>(a), static_cast<long long>(b),
+            static_cast<long long>(c), static_cast<long long>(d),
+            static_cast<long long>(e), static_cast<long long>(f));
+    LOGI("logi   %lld %lld %lld %lld %lld %lld",
+            static_cast<long long>(a), static_cast<long long>(b),
+            static_cast<long long>(c), static_cast<long long>(d),
+            static_cast<long long>(e), static_cast<long long>(f));
+    LOGW("logw   %lld %lld %lld %lld %lld %lld",
+            static_cast<long long>(a), static_cast<long long>(b),
+            static_cast<long long>(c), static_cast<long long>(d),
+            static_cast<long long>(e), static_cast<long long>(f));
+    LOGE("loge   %lld %lld %lld %lld %lld %lld",
+            static_cast<long long>(a), static_cast<long long>(b),
+            static_cast<long long>(c), static_cast<long long>(d),
+            static_cast<long long>(e), static_cast<long long>(f));
 
     Log::Instance()->setLogLevel(level);
 }
-
